+++ scenes/__init__.py
@@ -3,6 +3,7 @@ from .main_menu import MainMenuScene
 from .game_scene import GameScene
 from .pause_menu import PauseMenuScene
 from .game_over import GameOverScene
+from .options_menu import OptionsMenuScene
 
 __all__ = [
     "Scene",
@@ -10,4 +11,5 @@ __all__ = [
     "GameScene",
     "PauseMenuScene",
     "GameOverScene",
+    "OptionsMenuScene",
 ]